agent/requirements.txt
livekit-agents[deepgram,openai,elevenlabs]>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
